Moved dispatcher related definitions into InMemoryLedgerReader.
