+++ ledger/ledger-on-memory/src/main/scala/com/daml/ledger/on/memory/InMemoryBatchedLedgerReaderWriter.scala
@@ -12,7 +12,7 @@ import akka.stream.scaladsl.Source
 import com.daml.api.util.TimeProvider
 import com.daml.caching.Cache
 import com.daml.ledger.api.health.{HealthStatus, Healthy}
-import com.daml.ledger.on.memory.InMemoryLedgerReaderWriter.{DefaultTimeProvider, Index, dispatcher}
+import com.daml.ledger.on.memory.InMemoryLedgerReaderWriter.Index
 import com.daml.ledger.participant.state.kvutils.DamlKvutils.DamlStateValue
 import com.daml.ledger.participant.state.kvutils.api.{LedgerReader, LedgerRecord, LedgerWriter}
 import com.daml.ledger.participant.state.kvutils.{Bytes, KeyValueCommitting}
@@ -75,6 +75,8 @@ final class InMemoryBatchedLedgerReaderWriter(
 }
 
 object InMemoryBatchedLedgerReaderWriter {
+  val DefaultTimeProvider: TimeProvider = TimeProvider.UTC
+
   final class SingleParticipantOwner(
       initialLedgerId: Option[LedgerId],
       participantId: ParticipantId,
@@ -89,7 +91,7 @@ object InMemoryBatchedLedgerReaderWriter {
     ): Resource[InMemoryBatchedLedgerReaderWriter] = {
       val state = InMemoryState.empty
       for {
-        dispatcher <- dispatcher.acquire()
+        dispatcher <- InMemoryLedgerReader.dispatcher.acquire()
         readerWriter <- new Owner(
           initialLedgerId,
           participantId,
